requirements.txt
aiohttp
aioredis
astunparse
async-timeout
attrs
braceexpand
chardet
git+https://github.com/Rapptz/discord.py
fuzzywuzzy
humanize
idna
import-expression
jishaku
multidict
python-Levenshtein
six
typing-extensions
watchdog
yarl
pyparsing
discord-ext-colours
orjson
dnspython 
emoji
wand